Combine room and active room
