+++ client/src/Containers/App/styles.css
@@ -2,7 +2,7 @@ body {
   margin: 0;
   padding: 0;
   font-family: sans-serif;
-  overflow-y: hidden;
+  overflow-y: auto;
   height: 100%;
   display: flex;
 }
