+++ client/src/Redux/Reducers/initialState.js
@@ -11,70 +11,69 @@ export default {
   chat: {
     connected: false
   },
-
-  activeRoom: {
-    activeRoom: null
-  },
   
-  rooms: [{ 
-    roomID: 1,
-    sessionID: 0, // Latest session ID
-    humans: {
-        counsellor: 1,
-        user: 11,
+  room: {
+    activeRoom: null,
+    rooms: [{ 
+      roomID: 1,
+      sessionID: 0, // Latest session ID
+      humans: {
+          counsellor: 1,
+          user: 11,
+        },
+       messages: [
+       {
+        ID: 1,
+        sessionID: 1,
+        messageTime: 100,
+        counsellorID: 1,
+        userID: 1,
+        messageContent: 'Welcome to React Chat -- Built using React, Redux, Express, and Socket.io',
+        fromCounsellor: 1,
+        fromTwilio: 0
+       },
+       {
+        ID: 2,
+        sessionID: 2,
+        messageTime: 200,
+        counsellorID: 1,
+        userID: 1,
+        message: 'Wait, what are these words?',
+        fromCounsellor: 0,
+        fromTwilio: 0
+       }
+      ] },
+      {
+        roomID: 2,
+        sessionID: 1,
+        humans: {
+        counsellor: 2,
+        user: 12
       },
-     messages: [
-     {
-      ID: 1,
-      sessionID: 1,
-      messageTime: 100,
-      counsellorID: 1,
-      userID: 1,
-      messageContent: 'Welcome to React Chat -- Built using React, Redux, Express, and Socket.io',
-      fromCounsellor: 1,
-      fromTwilio: 0
+       messages: [
+       {
+        ID: 3,
+        sessionID: 3,
+        messageTime: 100,
+        counsellorID: 2,
+        userID: 12,
+       messageContent: 'Someone explain Sockets to me Please',
+       fromCounsellor: 1,
+       fromTwilio: 0
      },
      {
-      ID: 2,
-      sessionID: 2,
-      messageTime: 200,
-      counsellorID: 1,
-      userID: 1,
-      message: 'Wait, what are these words?',
-      fromCounsellor: 0,
-      fromTwilio: 0
-     }
-    ] },
-    {
-      roomID: 2,
-      sessionID: 1,
-      humans: {
-      counsellor: 2,
-      user: 12
-    },
-     messages: [
-     {
-      ID: 3,
+      ID: 4,
       sessionID: 3,
-      messageTime: 100,
+      messageTime: 200,
       counsellorID: 2,
       userID: 12,
-     messageContent: 'Someone explain Sockets to me Please',
-     fromCounsellor: 1,
-     fromTwilio: 0
-   },
-   {
-    ID: 4,
-    sessionID: 3,
-    messageTime: 200,
-    counsellorID: 2,
-    userID: 12,
-    messageContent: 'I have no idea',
-    fromCounsellor: 0,
-    fromTwilio: 0
-  }] 
-    }
-  ],
+      messageContent: 'I have no idea',
+      fromCounsellor: 0,
+      fromTwilio: 0
+    }] 
+      }
+    ],
+  },
 
   user: {
     user: {
