--- client/src/Redux/Types/activeRoomTypes.js
@@ -1,2 +1,0 @@
-export const SET_ACTIVE_ROOM = "set_active_room";
-export const REMOVE_ACTIVE_ROOM = "remove_active_room";
